Consolidate pytest fixtures into one file.
